Default empty string for targetValue
The SizeValidator does not allow you to do "0..{max}" constraints. When you populate a bean (using ORM for example) that has a constraint example of size "0..50", and the actual value in the database is NULL, the targetValue in the function is shown as "undefined". This leads to the validation always not passing and returning the error for required size. A default of "" in the arguments, not only allows the validation to properly work, but it also allows you to do stuff like:
"address"  : {
     type = "string",
     size = "0..50"
}