🐛 FIX: Correct build:docs script bugs
An error when running box run-script build:docs because variables.exportsDir doesn't exist.
The directoryCreate( outputDir ) is worse than useless, since DocBox creates this dir for you, and this code creates the wrong directory anyway! (it creates .tmp/apidocs/ inside the build/ directory, which is not where it should be.)